Define projections of cofeT, cmraT and ucmraT as Hint Extern.
This way, we can use eapply instead of class_apply, which is used
when the instances are defined using the Instance command. It seems
that eapply is stronger as class_apply, and as such solves some issues
when canonical structures have type class parameters, for example:

  Goal Op (option (dec_agree nat)). apply _.

This failed, but is fixed by this commit.