Remove the requirement that the unit of a CMRA is timeless.
This requirement was useful in Iris 2.0: in order to ensure that ownership of
the physical state was timeless, we required the ghost CMRA to have a timeless
unit. To avoid having additional type class parameters, or having to extend the
algebraic hierarchy, we required the units of any CMRA to be timeless.

In Iris 3.0, this issue no longer applies: ownership of the physical state is
ghost ownership in the global CMRA, whose unit is always timeless.

Thanks to Jeehoon Kang for spotting this unnecessary requirement.